#26: Rename as function_data; fix label bug
