+++ goalie/go_mesh_seq.py
@@ -4,8 +4,8 @@ Drivers for goal-oriented error estimation on sequences of meshes.
 
 from .adjoint import AdjointMeshSeq
 from .error_estimation import get_dwr_indicator
+from .function_data import IndicatorData
 from .log import pyrint
-from .solutions import IndicatorData
 from .utility import AttrDict
 from firedrake import Function, FunctionSpace, MeshHierarchy, TransferManager, project
 from firedrake.petsc import PETSc
