+++ goalie/mesh_seq.py
@@ -7,11 +7,11 @@ from firedrake.adjoint import pyadjoint
 from firedrake.adjoint_utils.solving import get_solve_blocks
 from firedrake.petsc import PETSc
 from firedrake.pyplot import triplot
+from .function_data import ForwardSolutionData
 from .interpolation import project
 from .log import pyrint, debug, warning, info, logger, DEBUG
 from .options import AdaptParameters
 from animate.quality import QualityMeasure
-from .solutions import ForwardSolutionData
 from .time_partition import TimePartition
 from .utility import AttrDict, Mesh
 from collections import OrderedDict
